ByteSpan list now sorted, so overlap happens less
Before, the bytespans would constantly overlap or be out of order. The
effect was that it was hard for the program to lay span tags out over
the NPs, because in order to do so, it would have to compensate for
already-placed tags. Now, the tags are in order, so this happens a LOT
less.