Implemented better algorithm for writing overlay
Previously, we would write per bytespan; that is, we would pull up a
bytespan and then write everything in that bytespan. This would happen
until the file ended. Now we walk through raw.txt and write a character
until we come to a bytespan opening. At that point, we create the open
tag.

The next step is to close these open tags at the appropriate time.