Implement a simple and cheap way of nesting the NPs accuartely
I just wanted to say, after spending so much time writing the code to
find the level of nesting each NP has, it is INFURIATING to find a
solution that is literally 1 line of code, which is what happened here.

The main difference is that I figured out how to use the CSS to properly
space the tags from the sides of the tags they are inside.