Fix JS function that marks up the coref chains
In the current implementation, when you click on any nested span, it
does the onclick operation for the inner tag, and then immediately does
the operation for the outer tag, which overwrites the inner tag's
result.

We fixed this by using a simple boolean flash that "resets" after a few
milliseconds.