Add stack to track click ids
Basically, in web development, if you click on a div or span tag that is
nested inside another one, then those onclick elements are activated
each in turn starting with the innermost tag you've clicked.

In our implementation, when you click on a span tag, we wait for a few
milliseconds to stop our changes from getting overwritten by outer tags.
As we wait, the ids that make this request are added to a stack, so that
we can use them to mark up the tracking div later.