Add link between coref elements in different analyses
The main idea of this abstraction is to give us a basis to allow users
to click on one coref element in one chain and have highlighting happen
in two different analyses of the document (say, our tool vs the gold
standard). We begin to enable this by allowing each coref element to
link to another one.