Add lookback for NP dependency adding.
To reiterate, the point is to be able to compare two different
interpretations of a document; we want to accomplish this by allowing
users to click on an NP to see what else is in the coref chain for both
interpretations of the document.

This commit looks to see if the current NP is a part of a larger NP that
began earlier.

It is critical that we keep track as we go, rather than simply look
backwards over the document. This is because an NP that contains the
current NP might be arbitrarily far back, and we would need to reparse the
whole document from that point back to guarantee that we got all nested
NPs. By keeping track of where NPs end as wo go along, we can just look
at a stack and see which NPs still haven't ended. It's just a better way
of doing it.