Add forwards-looking to NP association code
See previous commits for a more detailed explanation of what this code
does; basically we are finding which NPs overlap with other NPs in two
interpretations of a document, and we associate them by enumerating
their coref ids.

This commit solidified the lookback and implements the look forward
parts of this algorithm.