Add major fix to assoc coref id
In adding associated coref ids to the list, we tend to omit the first
one. This turns out to be a simple bookkeeping question; by setting one
of the indices to 1 for the initial case, and by changing the inside
loop to evaluate on 1 less, we have fixed the issue.