Add fixes for how we put tracking doc's coref ids in the html
There is a critical error in the code: when writing the opening html
tags for the tracking document's coref ids, we would use the coref id as
the index for the array that holds all the bytespan objects. This is
very clearly wrong; we want to use a normal counting index and then make
the span tag's class name the coref id of the bytespan object at that
index.

This commit introduces this (simple) fix.