Make the highlighting the atomic operation
When you click on an NP that is nested in another NP, the browser will
cue up the actions (I think they're smallest to largest). This causes
problems, because you will highlight one NP's coref chain, and then
another.

The solution (the best so far, anyway) is to wait a bit before we allow
any other NPs to be highlighted.