Begin to add provisions for resetting the assoc coref highlighting
Currently, when we click on an NP and its associated NP and their
chains get highlighted, we have to refresh before we click on another
NP. The reason is, if we don't, the last chain highlights directly over
the current one, which is terrible and blatantly wrong.

Here we add provisions to start fixing this by adding variables that
keep track of which ones were done last round.