Add method that builds the assoc coref id bit vector
This method builds a bit vector with a total size equal to the number of
bytes in the raw.txt file. At each particular element of this file,
there are a list of coref ids that would encapsulate this byte if were
were in the raw.txt file instead of the array.

These coref ids are split into two seperate arrays: one for the first
interpretation and one for the second. So, in other words, each element
in the array is an array that contains 2 arrays. Each of these contains
all the coref ids that the corresponding byte in raw.txt would be
inside.