Change awkward <br> handling inside <span>
The current implementation adds <br> tags wherever there is a newline
character in the original source document. Since we are drawing boxes
around NPs, if there is a newline somewhere inside an NP, we end up with
a weird-looking two-line box that represents the NP.

It's hard to describe, but what you need to know is that this commit
basically makes all NPs display display themselves in a one-line box (as
opposed to a two-line box.